+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+* [PR #502] Bump react-native-tab-view version
 * [PR #500] Search slips using the user geo location when not logged in
 * [PR #498] Use api call to action text
 * [PR #496] Show sent email when recovering password